ai/api/requirements.txt
Django>=3.2,<4.0
Pillow>=9.3.0
-f https://download.pytorch.org/whl/torch_stable.html
torch>=1.0.0+cpu
torchvision>=0.8.0+cpu
django-cors-headers